[REFACTOR/#38] usecase 패키지 분리
