+++ app/src/main/java/com/growthook/aos/presentation/cavecreate/CreateNewCaveViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.growthook.aos.domain.usecase.PostCaveUseCase
+import com.growthook.aos.domain.usecase.cavedetail.PostCaveUseCase
 import com.growthook.aos.domain.usecase.local.GetUserUseCase
 import com.growthook.aos.util.extension.addSourceList
 import dagger.hilt.android.lifecycle.HiltViewModel
